+++ client/cody/src/completions/providers/provider.ts
@@ -3,8 +3,9 @@ import { ReferenceSnippet } from '../context'
 
 export interface ProviderConfig {
     /**
-     * A factory to create instances of the provider. This pattern allows us to inject provider
-     * specific parameters outside of the callers of the factory.
+     * A factory to create instances of the provider. This pattern allows us to
+     * inject provider specific parameters outside of the callers of the
+     * factory.
      */
     create(options: ProviderOptions): Provider
 
@@ -19,6 +20,11 @@ export interface ProviderConfig {
     maximumContextCharacters: number
 
     /**
+     * When set, multi-line completions will trigger more often. This is
+     */
+    enableExtendedMultilineTriggers: boolean
+
+    /**
      * A string identifier used in event logs
      */
     identifier: string
@@ -29,7 +35,6 @@ export interface ProviderOptions {
     suffix: string
     fileName: string
     languageId: string
-    snippets: ReferenceSnippet[]
     multilineMode: null | 'block'
     // Relative length to `maximumContextCharacters`
     responsePercentage: number
@@ -44,7 +49,6 @@ export abstract class Provider {
     protected suffix: string
     protected fileName: string
     protected languageId: string
-    protected snippets: ReferenceSnippet[]
     protected multilineMode: null | 'block'
     protected responsePercentage: number
     protected prefixPercentage: number
@@ -56,7 +60,6 @@ export abstract class Provider {
         suffix,
         fileName,
         languageId,
-        snippets,
         multilineMode,
         responsePercentage,
         prefixPercentage,
@@ -67,7 +70,6 @@ export abstract class Provider {
         this.suffix = suffix
         this.fileName = fileName
         this.languageId = languageId
-        this.snippets = snippets
         this.multilineMode = multilineMode
         this.responsePercentage = responsePercentage
         this.prefixPercentage = prefixPercentage
@@ -75,5 +77,5 @@ export abstract class Provider {
         this.n = n
     }
 
-    public abstract generateCompletions(abortSignal: AbortSignal): Promise<Completion[]>
+    public abstract generateCompletions(abortSignal: AbortSignal, snippets: ReferenceSnippet[]): Promise<Completion[]>
 }