+++ client/web/src/cody/components/ChatUI/ChatUi.tsx
@@ -57,6 +57,7 @@ export const ChatUI: React.FC<IChatUIProps> = ({ codyChatStore, isSourcegraphApp
         toggleIncludeInferredRepository,
         toggleIncludeInferredFile,
         abortMessageInProgress,
+        fetchRepositoryNames,
     } = codyChatStore
 
     const [formInput, setFormInput] = useState('')
@@ -77,8 +78,22 @@ export const ChatUI: React.FC<IChatUIProps> = ({ codyChatStore, isSourcegraphApp
     const onEdit = useCallback((text: string) => editMessage(text), [editMessage])
 
     const scopeSelectorProps = useMemo(
-        () => ({ scope, setScope, toggleIncludeInferredRepository, toggleIncludeInferredFile }),
-        [scope, setScope, toggleIncludeInferredRepository, toggleIncludeInferredFile]
+        () => ({
+            scope,
+            setScope,
+            toggleIncludeInferredRepository,
+            toggleIncludeInferredFile,
+            fetchRepositoryNames,
+            isSourcegraphApp,
+        }),
+        [
+            scope,
+            setScope,
+            toggleIncludeInferredRepository,
+            toggleIncludeInferredFile,
+            fetchRepositoryNames,
+            isSourcegraphApp,
+        ]
     )
 
     if (!loaded) {