+++ client/web/src/enterprise/app/setup/steps/AppAllSetSetupStep.tsx
@@ -1,15 +1,21 @@
-import { FC, useContext } from 'react'
+import { FC, useCallback, useContext, useState } from 'react'
 
 import classNames from 'classnames'
 
-import { Button, H1, Text } from '@sourcegraph/wildcard'
+import { Button, H1, Text, Tooltip } from '@sourcegraph/wildcard'
 
-import { SetupStepsContext, StepComponentProps } from '../../../../setup-wizard/components'
+import { FooterWidget, SetupStepsContext, StepComponentProps } from '../../../../setup-wizard/components'
+import { AppSetupProgressBar } from '../components/AppSetupProgressBar'
 
 import styles from './AppAllSetSetupStep.module.scss'
 
 export const AppAllSetSetupStep: FC<StepComponentProps> = ({ className }) => {
     const { onNextStep } = useContext(SetupStepsContext)
+    const [isProgressFinished, setProgressFinished] = useState(false)
+
+    const handleOneRepositoryFinished = useCallback(() => {
+        setProgressFinished(true)
+    }, [])
 
     return (
         <div className={classNames(styles.root, className)}>
@@ -18,16 +24,35 @@ export const AppAllSetSetupStep: FC<StepComponentProps> = ({ className }) => {
 
                 <div className={styles.descriptionContent}>
                     <Text className={styles.descriptionText}>
-                        Open the app to get started. You can also access Cody from the system tray to chat with Cody
-                        alongside your editor.
+                        Once embeddings are finished being generated, you can specify Cody’s context and start asking
+                        questions in the Cody Chat.
                     </Text>
-                    <Button size="lg" variant="primary" className={styles.descriptionButton} onClick={onNextStep}>
-                        Open the app →
-                    </Button>
+
+                    <Tooltip content={!isProgressFinished ? 'Embeddings are still being generated' : ''}>
+                        <Button
+                            size="lg"
+                            variant="primary"
+                            disabled={!isProgressFinished}
+                            className={styles.descriptionButton}
+                            onClick={onNextStep}
+                        >
+                            Get Started
+                        </Button>
+                    </Tooltip>
                 </div>
             </div>
 
-            <img src="https://storage.googleapis.com/sourcegraph-assets/all-set.png" alt="" className={styles.image} />
+            <div className={styles.imageWrapper}>
+                <img
+                    src="https://storage.googleapis.com/sourcegraph-assets/setup/vscode-and-cody-chat.png"
+                    alt=""
+                    className={styles.image}
+                />
+            </div>
+
+            <FooterWidget>
+                <AppSetupProgressBar onOneRepositoryFinished={handleOneRepositoryFinished} className={styles.footer} />
+            </FooterWidget>
         </div>
     )
 }