+++ client/web/src/repo/blob/own/FileOwnershipPanel.tsx
@@ -82,8 +82,17 @@ export const FileOwnershipPanel: React.FunctionComponent<OwnershipPanelProps & T
                 filePath={filePath}
                 refetch={refetch}
                 showAddOwnerButton={true}
+                canAssignOwners={canAssignOwners}
             />
         )
     }
-    return <OwnerList filePath={filePath} repoID={repoID} refetch={refetch} showAddOwnerButton={true} />
+    return (
+        <OwnerList
+            filePath={filePath}
+            repoID={repoID}
+            refetch={refetch}
+            showAddOwnerButton={true}
+            canAssignOwners={canAssignOwners}
+        />
+    )
 }