+++ enterprise/internal/batches/state/state.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/batches/sources/azuredevops"
 	gerritbatches "github.com/sourcegraph/sourcegraph/enterprise/internal/batches/sources/gerrit"
+	"github.com/sourcegraph/sourcegraph/internal/extsvc"
 	adobatches "github.com/sourcegraph/sourcegraph/internal/extsvc/azuredevops"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/gerrit"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver/protocol"
@@ -24,7 +25,6 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/authz"
 	"github.com/sourcegraph/sourcegraph/internal/database"
-	"github.com/sourcegraph/sourcegraph/internal/extsvc"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/bitbucketcloud"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/bitbucketserver"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/github"
@@ -66,6 +66,7 @@ func SetDerivedState(ctx context.Context, repoStore database.RepoStore, client g
 	} else {
 		c.ExternalState = state
 	}
+
 	if state, err := computeReviewState(c, history); err != nil {
 		logger.Warn("Computing changeset review state", log.Error(err))
 	} else {
@@ -121,8 +122,10 @@ func computeCheckState(c *btypes.Changeset, events ChangesetEvents) btypes.Chang
 		return computeBitbucketCloudBuildState(c.UpdatedAt, m, events)
 	case *azuredevops.AnnotatedPullRequest:
 		return computeAzureDevOpsBuildState(m)
-	case *gerritbatches.AnnotatedChange, *protocol.PerforceChangelistState:
-		// Gerrit and Perforce don't have builds built-in, its better to be explicit by still
+	case *gerritbatches.AnnotatedChange:
+		return computeGerritBuildState(m)
+	case *protocol.PerforceChangelistState:
+		// Perforce doesn't have builds built-in, its better to be explicit by still
 		// including this case for clarity.
 		return btypes.ChangesetCheckStateUnknown
 	}
@@ -146,20 +149,37 @@ func computeExternalState(c *btypes.Changeset, history []changesetStatesAtTime, 
 	return newestDataPoint.externalState, nil
 }
 
+// computeSingleChangesetExternalState computes the reviewState for a github changeset
+func computeGitHubReviewState(c *btypes.Changeset) btypes.ChangesetReviewState {
+
+	// GitHub only stores the ReviewDecision in PullRequest metadata, not
+	// in events, so we need to handle it separtely. We want to respect the
+	// CODEOWNERS review as the mergeable state, not any other approval.
+	switch c.Metadata.(*github.PullRequest).ReviewDecision {
+	case "REVIEW_REQUIRED":
+		return btypes.ChangesetReviewStatePending
+	case "APPROVED":
+		return btypes.ChangesetReviewStateApproved
+	case "CHANGES_REQUESTED":
+		return btypes.ChangesetReviewStateChangesRequested
+	default:
+		return btypes.ChangesetReviewStatePending
+	}
+}
+
 // computeReviewState computes the review state for the changeset and its
 // associated events. The events should be presorted.
 func computeReviewState(c *btypes.Changeset, history []changesetStatesAtTime) (btypes.ChangesetReviewState, error) {
+
+	if c.ExternalServiceType == extsvc.TypeGitHub {
+		return computeGitHubReviewState(c), nil
+	}
+
 	if len(history) == 0 {
 		return computeSingleChangesetReviewState(c)
 	}
 
 	newestDataPoint := history[len(history)-1]
-
-	// GitHub only stores the ReviewState in events, we can't look at the
-	// Changeset.
-	if c.ExternalServiceType == extsvc.TypeGitHub {
-		return newestDataPoint.reviewState, nil
-	}
 
 	// For other codehosts we check whether the Changeset is newer or the
 	// events and use the newest entity to get the reviewstate.
@@ -218,6 +238,7 @@ func parseBitbucketServerBuildState(s string) btypes.ChangesetCheckState {
 	default:
 		return btypes.ChangesetCheckStateUnknown
 	}
+
 }
 
 func computeBitbucketCloudBuildState(lastSynced time.Time, apr *bbcs.AnnotatedPullRequest, events []*btypes.ChangesetEvent) btypes.ChangesetCheckState {
@@ -286,6 +307,38 @@ func parseAzureDevOpsBuildState(s adobatches.PullRequestStatusState) btypes.Chan
 	case adobatches.PullRequestBuildStatusStatePending:
 		return btypes.ChangesetCheckStatePending
 	case adobatches.PullRequestBuildStatusStateSucceeded:
+		return btypes.ChangesetCheckStatePassed
+	default:
+		return btypes.ChangesetCheckStateUnknown
+	}
+}
+
+func computeGerritBuildState(ac *gerritbatches.AnnotatedChange) btypes.ChangesetCheckState {
+	stateMap := make(map[string]btypes.ChangesetCheckState)
+
+	// States from last sync.
+	for _, reviewer := range ac.Reviewers {
+		for key, val := range reviewer.Approvals {
+			if key != gerrit.CodeReviewKey {
+				stateMap[key] = parseGerritBuildState(val)
+			}
+		}
+	}
+
+	states := make([]btypes.ChangesetCheckState, 0, len(stateMap))
+	for _, v := range stateMap {
+		states = append(states, v)
+	}
+	return combineCheckStates(states)
+}
+
+func parseGerritBuildState(s string) btypes.ChangesetCheckState {
+	switch s {
+	case "-2", "-1":
+		return btypes.ChangesetCheckStateFailed
+	case " 0":
+		return btypes.ChangesetCheckStatePending
+	case "+2", "+1":
 		return btypes.ChangesetCheckStatePassed
 	default:
 		return btypes.ChangesetCheckStateUnknown
@@ -609,18 +662,11 @@ func computeSingleChangesetExternalState(c *btypes.Changeset) (s btypes.Changese
 }
 
 // computeSingleChangesetReviewState computes the review state of a Changeset.
-// GitHub doesn't keep the review state on a changeset, so a GitHub Changeset
-// will always return ChangesetReviewStatePending.
-//
 // This method should NOT be called directly. Use computeReviewState instead.
 func computeSingleChangesetReviewState(c *btypes.Changeset) (s btypes.ChangesetReviewState, err error) {
 	states := map[btypes.ChangesetReviewState]bool{}
 
 	switch m := c.Metadata.(type) {
-	case *github.PullRequest:
-		// For GitHub we need to use `ChangesetEvents.ReviewState`.
-		return btypes.ChangesetReviewStatePending, nil
-
 	case *bitbucketserver.PullRequest:
 		for _, r := range m.Reviewers {
 			switch r.Status {
@@ -696,16 +742,21 @@ func computeSingleChangesetReviewState(c *btypes.Changeset) (s btypes.ChangesetR
 			//   0 : no score
 			//  -1 : needs changes
 			//  -2 : rejected
-			switch reviewer.Approvals.CodeReview {
-			case "+2", "+1":
-				states[btypes.ChangesetReviewStateApproved] = true
-			case " 0": // This isn't a typo, there is actually a space in the string.
-				states[btypes.ChangesetReviewStatePending] = true
-			case "-1", "-2":
-				states[btypes.ChangesetReviewStateChangesRequested] = true
-			default:
-				states[btypes.ChangesetReviewStatePending] = true
+			for key, val := range reviewer.Approvals {
+				if key == gerrit.CodeReviewKey {
+					switch val {
+					case "+2", "+1":
+						states[btypes.ChangesetReviewStateApproved] = true
+					case " 0": // This isn't a typo, there is actually a space in the string.
+						states[btypes.ChangesetReviewStatePending] = true
+					case "-1", "-2":
+						states[btypes.ChangesetReviewStateChangesRequested] = true
+					default:
+						states[btypes.ChangesetReviewStatePending] = true
+					}
+				}
 			}
+
 		}
 	case *protocol.PerforceChangelist:
 		states[btypes.ChangesetReviewStatePending] = true